refactor: rename flaskr to api
