+++ api/db/lists.py
@@ -1,3 +1,29 @@
+from os import environ
+
+import psycopg2
+import psycopg2.extras
+import json
+
+from dotenv import load_dotenv
+
+from dishes import add_dish
+
+load_dotenv()
+
+db = f"dbname={environ["DATABASE"]} host={environ["DATABASE_HOST"]} user={environ["DATABASE_USER"]} password={environ["DATABASE_PASSWORD"]} port={environ["DATABASE_PORT"]}"
+db_table = "testing_lists"
+
+
+# Get dishes list
+def get_shopping_lists(user):
+    with psycopg2.connect(db) as conn:
+        with conn.cursor(cursor_factory=psycopg2.extras.DictCursor) as curs:
+            curs.execute(f"SELECT * FROM {db_table} WHERE userid = %s;", (user["userid"],))
+            res = curs.fetchall()
+    
+    return res if res else "No lists"
+
+
 def create_list(list):
     with psycopg2.connect(db) as conn:
         with conn.cursor(cursor_factory=psycopg2.extras.DictCursor) as curs:
