+++ api/tests/lists/test_delete_list.py
