+++ api/tests/test_register.py
